renamed api to apiTypes in src/types
