+++ src/App.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
 import './App.css';
-import { HealthCheckResponse } from './types/api';
+import { HealthCheckResponse } from './types/apiTypes';
 import SignUpPage from './pages/SignUpPage';
 
 // Initialize React Query client
